2_preprocessing/requirements-freeze.txt
absl-py==2.1.0
anyio==4.8.0
astunparse==1.6.3
blosc2==3.0.0
cachetools==5.5.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cuda-python==12.6.0
cudf-cu12==24.12.0
cupy-cuda12x==13.3.0
cycler==0.12.1
dask==2025.1.0
dask-cudf-cu12==24.12.0
dask-expr==1.1.19
dask-glm==0.3.2
dask-ml==2024.4.4
dask-xgboost==0.2.0
distributed==2025.1.0
et-xmlfile==2.0.0
fastrlock==0.8.3
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.55.3
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
jinja2==3.1.4
joblib==1.4.2
keras==3.8.0
keras-tuner==1.4.7
kiwisolver==1.4.8
kt-legacy==1.0.5
libclang==18.1.1
libcudf-cu12==24.12.0
libkvikio-cu12==24.12.1
llvmlite==0.44.0
locket==1.0.0
markdown==3.7
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
multipledispatch==1.0.0
namex==0.0.8
natsort==8.4.0
ndindex==1.9.2
networkx==3.3
numba==0.61.0
numba-cuda==0.0.17.1
numexpr==2.10.2
numpy==1.26.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvcomp-cu12==4.1.0.6
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
nvtx==0.2.10
openpyxl==3.1.5
opt-einsum==3.4.0
optree==0.14.0
packaging==24.1
pandas==2.2.3
partd==1.4.2
pillow==10.2.0
protobuf==5.29.4
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==18.1.0
pygments==2.19.1
pylibcudf-cu12==24.12.0
pynvjitlink-cu12==0.4.0
pynvml==11.4.1
pyparsing==3.2.1
pyqt6==6.8.1
pyqt6-qt6==6.8.2
pyqt6-sip==13.10.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyyaml==6.0.2
rapids-dask-dependency==24.12.0
requests==2.32.3
rich==13.9.4
rmm-cu12==24.12.1
scapy==2.6.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==70.2.0
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
sortedcontainers==2.4.0
sparse==0.15.5
sympy==1.13.1
tables==3.10.2
tblib==3.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
threadpoolctl==3.5.0
toolz==1.0.0
torch==2.6.0+cu126
tornado==6.4.2
tqdm==4.67.1
triton==3.2.0
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xgboost==0.90
zict==3.0.0
